+++ src/main/java/com/softserve/demo/repository/ServicesRepository.java
@@ -3,5 +3,5 @@ package com.softserve.demo.repository;
 import com.softserve.demo.model.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicesRepository extends JpaRepository<Service, Long> {
+public interface ServicesRepository extends JpaRepository<Service, Integer> {
 }
